Analyse: split dataflow analysis module
Move the existing DFA implementation into a dedicated module while keeping the legacy analyse_dataflow path as a compatibility shim so downstream users keep working unchanged.
