+++ loki/analyse/__init__.py
@@ -9,3 +9,5 @@ Advanced analysis utilities, such as dataflow analysis functionalities.
 """
 
 from loki.analyse.analyse_dataflow import *  # noqa
+from loki.analyse.abstract_dfa import *  # noqa
+from loki.analyse.dataflow_analysis import *  # noqa
